backend/services/photos/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
alembic>=1.13.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
Pillow>=10.0.0
pillow-heif>=0.16.0
imagehash>=4.3.0
python-multipart>=0.0.7
httpx>=0.27.0
aio-pika>=9.0.0
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
cryptography>=41.0.0
